Split basic query search (exact match) from QuerySearchStrategy to its own BasicQuerySearchStrategy
